// src/lib/api/generated/models/UserBadgeWithBadge.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PolystirolHub Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Badge } from './Badge';
import { BadgeFromJSON, BadgeFromJSONTyped, BadgeToJSON } from './Badge';
import type { ExpiresAt } from './ExpiresAt';
import { ExpiresAtFromJSON, ExpiresAtFromJSONTyped, ExpiresAtToJSON } from './ExpiresAt';

/**
 *
 * @export
 * @interface UserBadgeWithBadge
 */
export interface UserBadgeWithBadge {
	/**
	 *
	 * @type {any}
	 * @memberof UserBadgeWithBadge
	 */
	id: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof UserBadgeWithBadge
	 */
	userId: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof UserBadgeWithBadge
	 */
	badgeId: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof UserBadgeWithBadge
	 */
	receivedAt: any | null;
	/**
	 *
	 * @type {ExpiresAt}
	 * @memberof UserBadgeWithBadge
	 */
	expiresAt?: ExpiresAt;
	/**
	 *
	 * @type {Badge}
	 * @memberof UserBadgeWithBadge
	 */
	badge: Badge;
}

/**
 * Check if a given object implements the UserBadgeWithBadge interface.
 */
export function instanceOfUserBadgeWithBadge(value: object): boolean {
	let isInstance = true;
	isInstance = isInstance && 'id' in value;
	isInstance = isInstance && 'userId' in value;
	isInstance = isInstance && 'badgeId' in value;
	isInstance = isInstance && 'receivedAt' in value;
	isInstance = isInstance && 'badge' in value;

	return isInstance;
}

export function UserBadgeWithBadgeFromJSON(json: any): UserBadgeWithBadge {
	return UserBadgeWithBadgeFromJSONTyped(json, false);
}

export function UserBadgeWithBadgeFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean
): UserBadgeWithBadge {
	if (json === undefined || json === null) {
		return json;
	}
	return {
		id: json['id'],
		userId: json['user_id'],
		badgeId: json['badge_id'],
		receivedAt: json['received_at'],
		expiresAt: !exists(json, 'expires_at') ? undefined : ExpiresAtFromJSON(json['expires_at']),
		badge: BadgeFromJSON(json['badge']),
	};
}

export function UserBadgeWithBadgeToJSON(value?: UserBadgeWithBadge | null): any {
	if (value === undefined) {
		return undefined;
	}
	if (value === null) {
		return null;
	}
	return {
		id: value.id,
		user_id: value.userId,
		badge_id: value.badgeId,
		received_at: value.receivedAt,
		expires_at: ExpiresAtToJSON(value.expiresAt),
		badge: BadgeToJSON(value.badge),
	};
}