// src/lib/api/generated/models/Badge.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PolystirolHub Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { BadgeType } from './BadgeType';
import { BadgeTypeFromJSON, BadgeTypeFromJSONTyped, BadgeTypeToJSON } from './BadgeType';
import type { Description } from './Description';
import { DescriptionFromJSON, DescriptionFromJSONTyped, DescriptionToJSON } from './Description';

/**
 *
 * @export
 * @interface Badge
 */
export interface Badge {
	/**
	 *
	 * @type {any}
	 * @memberof Badge
	 */
	name: any | null;
	/**
	 *
	 * @type {Description}
	 * @memberof Badge
	 */
	description?: Description;
	/**
	 *
	 * @type {BadgeType}
	 * @memberof Badge
	 */
	badgeType?: BadgeType;
	/**
	 *
	 * @type {any}
	 * @memberof Badge
	 */
	id: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof Badge
	 */
	imageUrl: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof Badge
	 */
	createdAt: any | null;
}

/**
 * Check if a given object implements the Badge interface.
 */
export function instanceOfBadge(value: object): boolean {
	let isInstance = true;
	isInstance = isInstance && 'name' in value;
	isInstance = isInstance && 'id' in value;
	isInstance = isInstance && 'imageUrl' in value;
	isInstance = isInstance && 'createdAt' in value;

	return isInstance;
}

export function BadgeFromJSON(json: any): Badge {
	return BadgeFromJSONTyped(json, false);
}

export function BadgeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Badge {
	if (json === undefined || json === null) {
		return json;
	}
	return {
		name: json['name'],
		description: !exists(json, 'description')
			? undefined
			: DescriptionFromJSON(json['description']),
		badgeType: !exists(json, 'badge_type') ? undefined : BadgeTypeFromJSON(json['badge_type']),
		id: json['id'],
		imageUrl: json['image_url'],
		createdAt: json['created_at'],
	};
}

export function BadgeToJSON(value?: Badge | null): any {
	if (value === undefined) {
		return undefined;
	}
	if (value === null) {
		return null;
	}
	return {
		name: value.name,
		description: DescriptionToJSON(value.description),
		badge_type: BadgeTypeToJSON(value.badgeType),
		id: value.id,
		image_url: value.imageUrl,
		created_at: value.createdAt,
	};
}